refactor(core/query-builder): extract qAlias/QEntity/proxy caches into query-builder/alias (#405)
Pure move; inline import() types replaced with a named ColumnJsonMeta import; re-exported from SelectQueryBuilder.ts.
